Removed message popup window in favor of main window highlighting
