+++ src/handlers/mod.rs
@@ -3,3 +3,4 @@ pub mod args;
 pub mod config;
 pub mod data;
 pub mod event;
+pub mod scroll;
